requirements.txt
python-telegram-bot==20.6
fpdf==1.7.2